requirements.txt
numpy==2.2.3
pillow==11.1.0
SerialPy==0.1.0